% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_form.R
\name{create_form}
\alias{create_form}
\title{Creates a Shiny Form 1st template}
\usage{
create_form(x)
}
\arguments{
\item{x}{The name of the form}
}
\value{

}
\description{
Creates a Shiny Form 1st template
}
\examples{
create_form("my form")
}
