% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iuform.R
\name{create_iuform}
\alias{create_iuform}
\title{Creates a Shiny Form 1st template}
\usage{
create_iuform(path, ...)
}
\arguments{
\item{x}{The name of the form}
}
\value{

}
\description{
Creates a Shiny Form 1st template
}
\examples{
create_form("my form")
}
