% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_end.R
\name{del_end}
\alias{del_end}
\title{delete the last char from some thing after counting its length}
\usage{
del_end(code)
}
\arguments{
\item{code}{this is code you want to trim its last char}
}
\description{
delete the last char from some thing after counting its length
}
