% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_word.R
\name{freq_word}
\alias{freq_word}
\title{Counts keywords in code}
\usage{
freq_word(code, pattern)
}
\arguments{
\item{code}{this is code you want to  search within}

\item{pattern}{this is the keywords you want to find}
}
\description{
Counts keywords in code
}
