% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{iuform: Indiana University Form}
\usage{
hello(x)
}
\arguments{
\item{x}{Please type your name}
}
\value{
the output from \code{\link{print}}
}
\description{
Developed at Indiana University- Masters in Biostatistics program to help create online data collection forms
}
\examples{
hello("Mo")
\dontrun{
hello("Indiana")
}
}
