% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_options.R
\name{menu_options}
\alias{menu_options}
\title{Menu Element options}
\usage{
menu_options(label, ...)
}
\arguments{
\item{label}{this is the tag of the menu element that you want to add options to}

\item{...}{the Menu options}
}
\description{
Menu Element options
}
