% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_button.R
\name{new_button}
\alias{new_button}
\title{form element}
\usage{
new_button(label, name, must_fill = F, section)
}
\arguments{
\item{label}{this is the tag of the form element}

\item{name}{the element name MUST be within quotes if more than one word}

\item{must_fill}{tags mandatory field}

\item{section}{the section where the element will be placed}
}
\description{
form element
}
