% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_checkbox.R
\name{new_checkbox}
\alias{new_checkbox}
\title{form element}
\usage{
new_checkbox(label, name, must_fill = F, section, Checked = F)
}
\arguments{
\item{label}{this is the ta of the form element}

\item{name}{the element name MUST be within quotes if more than one word}

\item{must_fill}{tags mandatory field}

\item{section}{the section where the element will be placed}

\item{Checked}{status of the check box at the start}
}
\description{
form element
}
