% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_menu.R
\name{new_menu}
\alias{new_menu}
\title{form element}
\usage{
new_menu(label, name, section)
}
\arguments{
\item{label}{this is the tag of the form element}

\item{name}{the element name MUST be within quotes if more than one word}

\item{section}{the section where the element will be placed}
}
\description{
form element
}
