% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housecleaning.R
\name{replace_word}
\alias{replace_word}
\title{Starting a new IUform Shiny aplication}
\usage{
replace_word(file, pattern, replace)
}
\arguments{
\item{file}{this is the file}

\item{pattern}{gives pattern}

\item{replace}{gives replacement}
}
\description{
Starting a new IUform Shiny aplication
}
