% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_options.R
\name{slider_options}
\alias{slider_options}
\title{Menu Element options}
\usage{
slider_options(label, ...)
}
\arguments{
\item{label}{this is the tag of the slider element that you want to add options to}

\item{...}{defines the lower number, upper number, and default number of slider}
}
\description{
Menu Element options
}
