% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iuform.R
\name{create_IUform}
\alias{create_IUform}
\title{Starting a new IUform Shiny aplication}
\usage{
create_IUform(path, open = TRUE, package_name = basename(path), author, ...)
}
\arguments{
\item{path}{this is the IUform name and the folder it would live in}

\item{open}{Boolean open the created project}

\item{package_name}{Default is `basename(path)` or this otherwise it will be `basename(getwd())`}

\item{author}{gives author name}

\item{...}{Arguments passed from the template.}
}
\description{
Starting a new IUform Shiny aplication
}
