% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_checkbox.R
\name{new_checkbox}
\alias{new_checkbox}
\title{form element}
\usage{
new_checkbox(label, must_fill = F, section)
}
\arguments{
\item{label}{this is the ta of the form element}

\item{must_fill}{tags mandatory field}

\item{section}{the section where the element will be placed}
}
\description{
form element
}
