% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_form.R
\name{new_form}
\alias{new_form}
\title{to start the form, ensures only one form}
\usage{
new_form(name, chaptered = F)
}
\arguments{
\item{name}{the  title/name of the form}

\item{chaptered}{if many section will be in the form}
}
\description{
to start the form, ensures only one form
}
