% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_freetext.R
\name{new_freetext}
\alias{new_freetext}
\title{form element}
\usage{
new_freetext(label, must_fill = F, section)
}
\arguments{
\item{label}{this is the tag of the form element}

\item{must_fill}{tags mandatory field}

\item{section}{the section where the element will be placed}
}
\description{
form element
}
